# luna-commons-template/src/main/java/com/luna/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "代码生成器",
    "termsOfService": "https://github.com/czy1024/luna-commons.git",
    "contact": {
      "email": "hexo.iszychen.club"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/scheduled": {
      "get": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "查询所有Scheduled",
        "description": "查询所有Scheduled方法详情",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "查询所有Scheduled",
            "schema": {
              "$ref": "#/definitions/Result«List«Scheduled»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "添加Scheduled",
        "description": "添加Scheduled方法详情",
        "operationId": "addUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "添加Scheduled方法详情",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Scheduled"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "添加Scheduled",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      }
    },
    "/scheduled/{id}": {
      "delete": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "根据ID删除Scheduled",
        "description": "根据ID删除Scheduled方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Scheduled方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "根据ID删除Scheduled",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "根据ID修改Scheduled",
        "description": "根据ID修改Scheduled方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Scheduled方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Scheduled",
            "description": "传入Scheduled的JSON对象",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Scheduled"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "根据ID修改Scheduled",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "根据ID查询Scheduled",
        "description": "根据ID查询Scheduled方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Scheduled方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "根据ID修改Scheduled",
            "schema": {
              "$ref": "#/definitions/Result«Scheduled»"
            }
          }
        }
      }
    },
    "/scheduled/search": {
      "post": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "不带分页条件搜索Scheduled",
        "description": "不带分页条件搜索Scheduled方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Scheduled",
            "description": "传入Scheduled的JSON对象",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Scheduled"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "不带分页搜索Scheduled",
            "schema": {
              "$ref": "#/definitions/Result«List«Scheduled»»"
            }
          }
        }
      }
    },
    "/scheduled/search/{page}/{size}": {
      "get": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "分页列表查询Scheduled",
        "description": "分页列表查询Scheduled方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "不带分页搜索Scheduled",
            "schema": {
              "$ref": "#/definitions/Result«List«Scheduled»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "ScheduledController"
        ],
        "summary": "分页条件搜索Scheduled",
        "description": "分页条件搜索Scheduled方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Scheduled",
            "description": "传入Scheduled的JSON对象",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Scheduled"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "分页条件搜索Scheduled",
            "schema": {
              "$ref": "#/definitions/Result«List«Scheduled»»"
            }
          }
        }
      }
    },
...
    "/template/{id}": {
      "delete": {
        "tags": [
          "TemplateController"
        ],
        "summary": "根据ID删除Template",
        "description": "根据ID删除Template方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Template方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "根据ID删除Template",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TemplateController"
        ],
        "summary": "根据ID修改Template",
        "description": "根据ID修改Template方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Template方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Template",
            "description": "传入Template的JSON对象",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Template"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "根据ID修改Template",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "TemplateController"
        ],
        "summary": "根据ID查询Template",
        "description": "根据ID查询Template方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Template方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "根据ID修改Template",
            "schema": {
              "$ref": "#/definitions/Result«Template»"
            }
          }
        }
      }
    },
    "/template/search": {
      "post": {
        "tags": [
          "TemplateController"
        ],
        "summary": "不带分页条件搜索Template",
        "description": "不带分页条件搜索Template方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Template",
            "description": "传入Template的JSON对象",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Template"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "不带分页搜索Template",
            "schema": {
              "$ref": "#/definitions/Result«List«Template»»"
            }
          }
        }
      }
    },
    "/template/search/{page}/{size}": {
      "get": {
        "tags": [
          "TemplateController"
        ],
        "summary": "分页列表查询Template",
        "description": "分页列表查询Template方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "不带分页搜索Template",
            "schema": {
              "$ref": "#/definitions/Result«List«Template»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TemplateController"
        ],
        "summary": "分页条件搜索Template",
        "description": "分页条件搜索Template方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Template",
            "description": "传入Template的JSON对象",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Template"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
          },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
          },
          "404": {
            "description": "not found(没有找到相关资源)"
          },
          "405": {
            "description": "Invalid input(无效的输入)"
          },
          "500": {
            "description": "服务器内部错误"
          },
          "200": {
            "description": "分页条件搜索Template",
            "schema": {
              "$ref": "#/definitions/Result«List«Template»»"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Result«List«Scheduled»»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "array",
          "description": "逻辑数据",
          "items": {
            "$ref": "#/definitions/Scheduled"
          }
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Result«Scheduled»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/Scheduled"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Scheduled": {
      "type": "object",
      "properties": {
        "cronId": {
          "type": "integer",
          "format": "int32",
          "description": "主键id"
        },
        "cronKey": {
          "type": "string",
          "description": "定时任务完整类名"
        },
        "cronExpression": {
          "type": "string",
          "description": "cron表达式"
        },
        "taskExplain": {
          "type": "string",
          "description": "任务描述"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态,1:正常;2:停用"
        }
      },
      "description": "Scheduled"
    },
    "Result«List«Template»»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "array",
          "description": "逻辑数据",
          "items": {
            "$ref": "#/definitions/Template"
          }
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Result«Template»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/Template"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Template": {
      "type": "object",
      "properties": {
        "id": {
          "type": "long",
          "description": "主键"
        },
        "createTime": {
          "type": "date",
          "description": "创建时间"
        },
        "modifiedTime": {
          "type": "date",
          "description": "修改时间"
        },
        "subject": {
          "type": "string",
          "description": "标题"
        },
        "content": {
          "type": "string",
          "description": "内容"
        }
      },
      "description": "Template"
    }
  },
...
}